chore: rework blast to use react-query, remove unused redux
